update to Float and add NormalEnemy

+++ app/src/main/java/com/ligrim/tower_defense/Enemy.java
@@ -3,13 +3,13 @@ package com.ligrim.tower_defense;
 public interface Enemy extends GameEntity {
 
     int getHealth();
-    double getSpeed();
+    float getSpeed();
     int getArmor();
     int getPrize();
     void reduceHealth(int damage);
-    void setDirection(double dx, double dy); // dx and dy is unit vector
-    double getDirectionX();
-    double getDirectionY();
+    void setDirection(float dx, float dy); // dx and dy is unit vector
+    float getDirectionX();
+    float getDirectionY();
 
 
 }
